requirements.txt
faiss-cpu
groq
langchain-groq
PyPDF2
langchain_google_genai
streamlit
langchain
langchain_community
python-dotenv
pypdf